+++ app/graphql/resolvers/concerns/work_items/look_ahead_preloads.rb
@@ -39,6 +39,7 @@ module WorkItems
         assignees: :assignees_by_name_and_id,
         participants: WorkItem.participant_includes,
         parent: :work_item_parent,
+        has_parent: :work_item_parent,
         children: { work_item_children_by_relative_position: [:author, { project: :project_feature }] },
         milestone: { milestone: [:project, :group] },
         subscribed: [:assignees, :award_emoji, { notes: [:author, :award_emoji] }],